levantamiento e algunas observaciones

+++ app/Http/Controllers/AsociadosController.php
@@ -115,7 +115,8 @@ class AsociadosController extends Controller
         //    var_dump(!empty($_REQUEST["idcargo"]));
         //    exit;
             //$_REQUEST["idcargo"] = (array) $_REQUEST["idcargo"];
-            if(isset($_REQUEST["idcargo"])) {
+            // echo gettype($_REQUEST["idcargo"]); exit;
+            if(isset($_REQUEST["idcargo"]) && gettype($_REQUEST["idcargo"]) == "array" && count($_REQUEST["idcargo"]) > 0) {
                 DB::table("iglesias.cargo_miembro")->where("idmiembro", $request->input("idmiembro"))->delete();
                 // print_r($this->preparar_datos("iglesias.cargo_miembro", $_POST, "D")); exit;
                 $result = $this->base_model->insertar($this->preparar_datos("iglesias.cargo_miembro", $_POST, "D"), "D");
@@ -124,7 +125,7 @@ class AsociadosController extends Controller
 
             //$_REQUEST["capacitacion"] = (array) $_REQUEST["capacitacion"];
 
-            if(isset($_REQUEST["capacitacion"])) {
+            if(isset($_REQUEST["capacitacion"]) && gettype($_REQUEST["capacitacion"]) == "array" && count($_REQUEST["capacitacion"]) > 0) {
                 DB::table("iglesias.capacitacion_miembro")->where("idmiembro", $request->input("idmiembro"))->delete();
                 // print_r($this->preparar_datos("iglesias.capacitacion_miembro", $_POST, "D")); exit;
                 $result = $this->base_model->insertar($this->preparar_datos("iglesias.capacitacion_miembro", $_POST, "D"), "D");
